Ensure binaries start, stop, and are cleaned up properly

+++ client.go
@@ -3,6 +3,7 @@ package main
 import (
 	"bufio"
 	"encoding/json"
+	"io"
 	"log"
 	"net"
 	"net/rpc"
@@ -14,50 +15,97 @@ import (
 	"github.com/jspc/loadtest"
 )
 
+const (
+	// RPCommand is the command to request from our RPC'd up scheduler
+	RPCCommand = "Server.Run"
+)
+
 type Job struct {
 	Users    int    `json:"users"`
 	Duration int64  `json:"duration"`
 	Binary   string `json:"binary"`
 
-	binaryPath string
+	bin        binary
 	items      int
 	process    *os.Process
 	connection net.Conn
+	setup      bool
+	complete   bool
+	service    *rpc.Client
 }
 
-func (j Job) Start() {
+func (j *Job) Start() {
 	go func() {
 		err := j.execute()
-		if err != nil {
+		if err != nil && err != io.EOF {
 			panic(err)
 		}
 	}()
 
-	connect := func() (err error) {
-		j.connection, err = net.Dial("tcp", loadtest.RPCAddr)
+	defer func() {
+		if j.process != nil {
+			j.process.Kill()
+			j.process.Wait()
+		}
+	}()
 
-		return
+	err := backoff.Retry(j.TryConnect, backoff.NewExponentialBackOff())
+	if err != nil {
+		panic(err)
 	}
-	err := backoff.Retry(connect, backoff.NewExponentialBackOff())
+	defer j.connection.Close()
 
-	service := rpc.NewClient(j.connection)
+	j.service = rpc.NewClient(j.connection)
+
+	err = backoff.Retry(j.TryRequest, backoff.NewExponentialBackOff())
+	if err != nil {
+		panic(err)
+	}
+	defer j.service.Close()
+
+	j.setup = true
 
 	go func() {
 		for {
-			err = service.Call("Server.Run", &loadtest.NullArg{}, &loadtest.NullArg{})
+			if j.complete {
+				return
+			}
+
+			err = j.TryRequest()
 			if err != nil {
-				log.Panic(err)
+				panic(err)
 			}
 		}
 	}()
 
 	time.Sleep(time.Duration(j.Duration) * time.Second)
 
-	j.process.Kill()
+	j.complete = true
+}
+
+func (j *Job) TryConnect() (err error) {
+	log.Print("try connect")
+	j.connection, err = net.Dial("tcp", loadtest.RPCAddr)
+
+	return
+}
+
+func (j *Job) TryRequest() (err error) {
+	if !j.setup {
+		log.Print("try request")
+	}
+
+	err = j.service.Call(RPCCommand, &loadtest.NullArg{}, &loadtest.NullArg{})
+
+	if err != nil && !j.complete {
+		return
+	}
+
+	return nil // Either err is nil, or we don't care because the command is over
 }
 
 func (j *Job) execute() (err error) {
-	cmd := exec.Command(j.binaryPath)
+	cmd := exec.Command(j.bin.Path)
 
 	stdout, err := cmd.StdoutPipe()
 	if err != nil {
@@ -66,7 +114,8 @@ func (j *Job) execute() (err error) {
 
 	rd := bufio.NewReader(stdout)
 
-	if err = cmd.Start(); err != nil {
+	err = cmd.Start()
+	if err != nil {
 		return
 	}
 
@@ -88,6 +137,7 @@ func (j *Job) execute() (err error) {
 		}
 
 		j.items++
+
 		// do something here with o
 	}
 }
